refactor: create the `UI.Event` module
