+++ app/UI.hs
@@ -40,10 +40,8 @@ import           Scene                      (SceneElement (WithSpeaker, WithoutS
 import           Talking                    (destruct, talkWith)
 import           UI.Attrs                   (attrMapForThisGame, emptyAttr,
                                              greenAttr, redAttr)
-
-data Tick = Tick
-
-type Name = ()
+import           UI.Event                   (handleEvent)
+import           UI.Types                   (Name, Tick (Tick))
 
 main :: IO ()
 main = do
@@ -63,37 +61,6 @@ app = App { appDraw = drawUI
           , appStartEvent = return
           , appAttrMap = const attrMapForThisGame
           }
-
-handleEvent :: Engine -> BrickEvent Name Tick -> EventM Name (Next Engine)
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey (V.KChar 'q') [])) = halt e
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey V.KEsc []))        = halt e
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey V.KUp []))         = handlePlayerMove (V2 0 1) e
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey V.KDown []))       = handlePlayerMove (V2 0 (-1)) e
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey V.KRight []))      = handlePlayerMove (V2 1 0) e
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey V.KLeft []))       = handlePlayerMove (V2 (-1) 0) e
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey V.KUpRight [])) = handlePlayerMove (V2 1 1) e
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey V.KUpLeft [])) = handlePlayerMove (V2 1 (-1)) e
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey V.KDownRight [])) = handlePlayerMove (V2 (-1) 1) e
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey V.KDownLeft [])) = handlePlayerMove (V2 (-1) (-1)) e
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey (V.KChar 'k') [])) = handlePlayerMove (V2 0 1) e
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey (V.KChar 'j') [])) = handlePlayerMove (V2 0 (-1)) e
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey (V.KChar 'l') [])) = handlePlayerMove (V2 1 0) e
-handleEvent e@PlayerIsExploring{} (VtyEvent (V.EvKey (V.KChar 'h') [])) = handlePlayerMove (V2 (-1) 0) e
-handleEvent e@HandlingScene{} (VtyEvent (V.EvKey V.KEnter [])) = handleMessageEvent e
-handleEvent e@Talking{} (VtyEvent (V.EvKey V.KEnter [])) = handleTalking e
-handleEvent e _                                     = continue e
-
-handlePlayerMove :: V2 Int -> Engine -> EventM Name (Next Engine)
-handlePlayerMove d e = continue $ flip execState e $ do
-    eng <- get
-    let finished = eng ^?! isGameOver
-    unless finished $ do
-        playerBumpAction d
-
-        eng <- get
-        case eng of
-            PlayerIsExploring {} -> completeThisTurn
-            _                    -> return ()
 
 drawUI :: Engine -> [Widget Name]
 drawUI e@PlayerIsExploring{} = [ C.center $ drawHpBar e <+> (padTop (Pad 2) (drawGame e) <=> drawMessageLog e)]
@@ -155,16 +122,3 @@ drawHpBar e = let barWidth = 20
                   filledWidth = currentHp * barWidth `div` maxHp
                   attrAt x = if x < filledWidth then greenAttr else redAttr
               in vBox [hBox $ map (\x -> withAttr (attrAt x) $ str "XX") [0 .. barWidth - 1], str $ "HP: " ++ show currentHp ++ " / " ++ show maxHp]
-
-handleMessageEvent :: Engine -> EventM Name (Next Engine)
-handleMessageEvent e@HandlingScene{} =
-        continue $ if length (e ^?! scene) > 1
-                    then e { _scene = tail (e ^?! scene) }
-                    else e ^?! afterFinish
-handleMessageEvent _ = error "unreachable"
-
-handleTalking :: Engine -> EventM Name (Next Engine)
-handleTalking e@Talking{} = continue $ do
-        let e' = e ^?! afterTalking
-        execState completeThisTurn e'
-handleTalking _ = error "unreachable"
