+++ app/UI/Types.hs
@@ -1,0 +1,8 @@
+module UI.Types
+    ( Tick(..)
+    , Name
+    ) where
+
+data Tick = Tick
+
+type Name = ()
